+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Apollo Workbench VSCode 3.2.12
 
-- Enable setting Router Version in VSCode settings
+- Enable setting Router Version in VSCode settings `routerVersion`
 - Better support `subgraph_url` to load read-only schema
 - Create `subgraph_url` watcher to ping changes
   - Introduce settings `enableSubgraphUrlWatcher` and `subgraphWatcherPingInterval` to allow a user to disable/enable the watcher and set the interval the watcher pings `subgraph_url` for updates.